s
s

+++ scripts/stop_server.sh
@@ -1,2 +1,5 @@
 #!/bin/bash
-sudo systemctl stop httpd
+isExistApp=`pgrep httpd`
+if [[ -n  $isExistApp ]]; then
+   sudo systemctl stop httpd
+fi
